@@ -9,14 +9,9 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Filtering
     /// Implements the logic of filtering out api.
     /// Reads the file with the list of attributes, types, members in DocId format.
     /// </summary>
-    public class DocIdSymbolFilter : ISymbolFilter
+    public class DocIdSymbolFilter(string[] docIdsToExcludeFiles) : ISymbolFilter
     {
-        private readonly HashSet<string> _docIdsToExclude;
-
-        public DocIdSymbolFilter(string[] docIdsToExcludeFiles)
-        {
-            _docIdsToExclude = new HashSet<string>(ReadDocIdsAttributes(docIdsToExcludeFiles));
-        }
+        private readonly HashSet<string> _docIdsToExclude = new(ReadDocIdsAttributes(docIdsToExcludeFiles));
 
         /// <summary>
         ///  Determines whether the <see cref="ISymbol"/> should be included.
@@ -34,13 +29,17 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Filtering
             return true;
         }
 
-        private static IEnumerable<string> ReadDocIdsAttributes(IEnumerable<string> excludeAttributesFiles)
+        private static IEnumerable<string> ReadDocIdsAttributes(IEnumerable<string> docIdsToExcludeFiles)
         {
-            foreach (string filePath in excludeAttributesFiles)
+            foreach (string docIdsToExcludeFile in docIdsToExcludeFiles)
             {
-                foreach (string id in File.ReadAllLines(filePath))
+                foreach (string id in File.ReadAllLines(docIdsToExcludeFile))
                 {
+#if NET
+                    if (!string.IsNullOrWhiteSpace(id) && !id.StartsWith('#') && !id.StartsWith("//"))
+#else
                     if (!string.IsNullOrWhiteSpace(id) && !id.StartsWith("#") && !id.StartsWith("//"))
+#endif
                     {
                         yield return id.Trim();
                     }
