@@ -196,7 +196,7 @@ namespace Microsoft.DotNet.ApiCompat.Tool
                 (string, string)[]? leftAssembliesTransformationPattern = parseResult.GetValue(leftAssembliesTransformationPatternOption);
                 (string, string)[]? rightAssembliesTransformationPattern = parseResult.GetValue(rightAssembliesTransformationPatternOption);
 
-                Func<ISuppressionEngine, SuppressibleConsoleLog> logFactory = (suppressionEngine) => new(suppressionEngine, verbosity);
+                SuppressibleConsoleLog logFactory(ISuppressionEngine suppressionEngine) => new(suppressionEngine, verbosity);
                 ValidateAssemblies.Run(logFactory,
                     generateSuppressionFile,
                     preserveUnnecessarySuppressions,
@@ -310,7 +310,7 @@ namespace Microsoft.DotNet.ApiCompat.Tool
                 Dictionary<NuGetFramework, IEnumerable<string>>? packageAssemblyReferences = parseResult.GetValue(packageAssemblyReferencesOption);
                 Dictionary<NuGetFramework, IEnumerable<string>>? baselinePackageAssemblyReferences = parseResult.GetValue(baselinePackageAssemblyReferencesOption);
 
-                Func<ISuppressionEngine, SuppressibleConsoleLog> logFactory = (suppressionEngine) => new(suppressionEngine, verbosity);
+                SuppressibleConsoleLog logFactory(ISuppressionEngine suppressionEngine) => new(suppressionEngine, verbosity);
                 ValidatePackage.Run(logFactory,
                     generateSuppressionFile,
                     preserveUnnecessarySuppressions,
@@ -341,29 +341,29 @@ namespace Microsoft.DotNet.ApiCompat.Tool
 
         private static string[][] ParseAssemblyReferenceArgument(ArgumentResult argumentResult)
         {
-            List<string[]> args = new();
+            List<string[]> args = [];
             foreach (var token in argumentResult.Tokens)
             {
                 args.Add(token.Value.Split(','));
             }
 
-            return args.ToArray();
+            return [.. args];
         }
 
         private static string[] ParseAssemblyArgument(ArgumentResult argumentResult)
         {
-            List<string> args = new();
+            List<string> args = [];
             foreach (var token in argumentResult.Tokens)
             {
                 args.AddRange(token.Value.Split(','));
             }
 
-            return args.ToArray();
+            return [.. args];
         }
 
         private static (string CaptureGroupPattern, string ReplacementString)[]? ParseTransformationPattern(ArgumentResult argumentResult)
         {
-            var patterns = new (string CaptureGroupPattern, string ReplacementPattern)[argumentResult.Tokens.Count];
+            var patterns = new(string CaptureGroupPattern, string ReplacementPattern)[argumentResult.Tokens.Count];
             for (int i = 0; i < argumentResult.Tokens.Count; i++)
             {
                 string[] parts = argumentResult.Tokens[i].Value.Split(';');
