@@ -9,31 +9,37 @@ namespace Microsoft.DotNet.ApiCompatibility
     /// <summary>
     /// Class representing a difference of compatibility, containing detailed information about it.
     /// </summary>
-    public readonly struct CompatDifference : IDiagnostic, IEquatable<CompatDifference>
+    /// <param name="left">The metadata information of the left comparison side.</param>
+    /// <param name="right">The metadata information of the right comparison side.</param>
+    /// <param name="diagnosticId"><see cref="string"/> representing the diagnostic ID.</param>
+    /// <param name="message"><see cref="string"/> message describing the difference.</param>
+    /// <param name="type"><see cref="DifferenceType"/> to describe the type of the difference.</param>
+    /// <param name="memberId"><see cref="string"/> containing the member ID for which the difference is associated to.</param>
+    public readonly struct CompatDifference(MetadataInformation left, MetadataInformation right, string diagnosticId, string message, DifferenceType type, string? memberId) : IDiagnostic, IEquatable<CompatDifference>
     {
         /// <inheritdoc />
-        public string DiagnosticId { get; }
+        public string DiagnosticId { get; } = diagnosticId;
 
         /// <summary>
         /// The <see cref="DifferenceType"/>.
         /// </summary>
-        public DifferenceType Type { get; }
+        public DifferenceType Type { get; } = type;
 
         /// <inheritdoc />
-        public string Message { get; }
+        public string Message { get; } = message;
 
         /// <inheritdoc />
-        public string? ReferenceId { get; }
+        public string? ReferenceId { get; } = memberId;
 
         /// <summary>
         ///  The left's metadata information.
         /// </summary>
-        public MetadataInformation Left { get; }
+        public MetadataInformation Left { get; } = left;
 
         /// <summary>
         /// The right's metadata information.
         /// </summary>
-        public MetadataInformation Right { get; }
+        public MetadataInformation Right { get; } = right;
 
         /// <summary>
         /// Instantiate a new object representing the compatibility difference.
@@ -50,25 +56,6 @@ namespace Microsoft.DotNet.ApiCompatibility
         }
 
         /// <summary>
-        /// Instantiate a new object representing the compatibility difference.
-        /// </summary>
-        /// <param name="left">The metadata information of the left comparison side.</param>
-        /// <param name="right">The metadata information of the right comparison side.</param>
-        /// <param name="diagnosticId"><see cref="string"/> representing the diagnostic ID.</param>
-        /// <param name="message"><see cref="string"/> message describing the difference.</param>
-        /// <param name="type"><see cref="DifferenceType"/> to describe the type of the difference.</param>
-        /// <param name="memberId"><see cref="string"/> containing the member ID for which the difference is associated to.</param>
-        public CompatDifference(MetadataInformation left, MetadataInformation right, string diagnosticId, string message, DifferenceType type, string? memberId)
-        {
-            Left = left;
-            Right = right;
-            DiagnosticId = diagnosticId;
-            Message = message;
-            Type = type;
-            ReferenceId = memberId;
-        }
-
-        /// <summary>
         /// Create a compatibility difference object with default left and right metadata for which the difference occurred.
         /// </summary>
         public static CompatDifference CreateWithDefaultMetadata(string diagnosticId, string message, DifferenceType type, string? memberId) =>
@@ -78,17 +65,6 @@ namespace Microsoft.DotNet.ApiCompatibility
                 message,
                 type,
                 memberId);
-
-        /// <summary>
-        /// Create a compatibility difference object with default left and right metadata for which the difference occurred.
-        /// </summary>
-        public static CompatDifference CreateWithDefaultMetadata(string diagnosticId, string message, DifferenceType type, ISymbol member) =>
-            new(MetadataInformation.DefaultLeft,
-                MetadataInformation.DefaultRight,
-                diagnosticId,
-                message,
-                type,
-                member);
 
         /// <summary>
         /// Gets a <see cref="string"/> representation of the difference.
