@@ -8,21 +8,14 @@ namespace Microsoft.DotNet.ApiCompatibility.Mapping
     /// <summary>
     /// Factory to create an ElementMapper instance.
     /// </summary>
-    public sealed class ElementMapperFactory : IElementMapperFactory
+    public sealed class ElementMapperFactory(IRuleRunner ruleRunner) : IElementMapperFactory
     {
-        private readonly IRuleRunner _ruleRunner;
-
-        public ElementMapperFactory(IRuleRunner ruleRunner)
-        {
-            _ruleRunner = ruleRunner;
-        }
-
         /// <inheritdoc />
         public IAssemblySetMapper CreateAssemblySetMapper(IMapperSettings settings, int rightCount) =>
-            new AssemblySetMapper(_ruleRunner, settings, rightCount);
+            new AssemblySetMapper(ruleRunner, settings, rightCount);
 
         /// <inheritdoc />
         public IAssemblyMapper CreateAssemblyMapper(IMapperSettings settings, int rightCount) =>
-            new AssemblyMapper(_ruleRunner, settings, rightCount);
+            new AssemblyMapper(ruleRunner, settings, rightCount);
     }
 }
