@@ -7,7 +7,11 @@ namespace Microsoft.DotNet.ApiCompatibility
     /// Struct containing the assembly's relevant information, used to distinguish different tuple comparisons
     /// and different list of <see cref="CompatDifference"/>.
     /// </summary>
-    public readonly struct MetadataInformation : IEquatable<MetadataInformation>
+    public readonly struct MetadataInformation(string assemblyName,
+        string assemblyId,
+        string? fullPath = null,
+        IEnumerable<string>? references = null,
+        string? displayString = null) : IEquatable<MetadataInformation>
     {
         private const string DEFAULT_LEFT_NAME = "left";
         private const string DEFAULT_RIGHT_NAME = "right";
@@ -25,36 +29,27 @@ namespace Microsoft.DotNet.ApiCompatibility
         /// <summary>
         /// The name of the assembly.
         /// </summary>
-        public readonly string AssemblyName;
+        public readonly string AssemblyName = assemblyName;
 
         /// <summary>
         /// The unique assembly id.
         /// </summary>
-        public readonly string AssemblyId;
+        public readonly string AssemblyId = assemblyId;
 
         /// <summary>
         /// Returns the assembly's full path or if it's part of an archive, the path to the archive.
         /// </summary>
-        public readonly string FullPath;
+        public readonly string FullPath = fullPath ?? assemblyId;
 
         /// <summary>
         /// The assembly references.
         /// </summary>
-        public readonly IEnumerable<string>? References;
+        public readonly IEnumerable<string>? References = references;
 
         /// <summary>
         /// The assembly's display string.
         /// </summary>
-        public readonly string DisplayString;
-
-        public MetadataInformation(string assemblyName, string assemblyId, string? fullPath = null, IEnumerable<string>? references = null, string? displayString = null)
-        {
-            AssemblyName = assemblyName;
-            AssemblyId = assemblyId;
-            FullPath = fullPath ?? assemblyId;
-            References = references;
-            DisplayString = displayString ?? assemblyId;
-        }
+        public readonly string DisplayString = displayString ?? assemblyId;
 
         /// <inheritdoc />
         public override bool Equals(object? obj) =>
@@ -69,11 +64,15 @@ namespace Microsoft.DotNet.ApiCompatibility
         /// <inheritdoc />
         public override int GetHashCode()
         {
+#if NET
+            return HashCode.Combine(AssemblyName, AssemblyId, FullPath);
+#else
             int hashCode = 1447485498;
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(AssemblyName);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(AssemblyId);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(FullPath);
             return hashCode;
+#endif
         }
 
         /// <inheritdoc />
