@@ -9,28 +9,19 @@ namespace Microsoft.DotNet.ApiCompatibility.Rules
     /// <summary>
     /// Rule runner that exposes functionality to initialize rules and run element mapper objects.
     /// </summary>
-    public class RuleRunner : IRuleRunner
+    public class RuleRunner(IRuleFactory ruleFactory, IRuleContext context) : IRuleRunner
     {
-        private readonly IRuleFactory _ruleFactory;
-        private readonly IRuleContext _context;
-
-        public RuleRunner(IRuleFactory ruleFactory, IRuleContext context)
-        {
-            _ruleFactory = ruleFactory;
-            _context = context;
-        }
-
         /// <inheritdoc />
         public void InitializeRules(IRuleSettings settings)
         {
             // Instantiate the rules but don't invoke anything on them as they register themselves on "events" inside their constructor.
-            _ = _ruleFactory.CreateRules(settings, _context);
+            _ = ruleFactory.CreateRules(settings, context);
         }
 
         /// <inheritdoc />
         public IEnumerable<CompatDifference> Run<T>(IElementMapper<T> mapper)
         {
-            List<CompatDifference> differences = new();
+            List<CompatDifference> differences = [];
 
             int rightLength = mapper.Right.Length;
             for (int rightIndex = 0; rightIndex < rightLength; rightIndex++)
@@ -46,7 +37,7 @@ namespace Microsoft.DotNet.ApiCompatibility.Rules
                        the assembly mapper is directly visited. */
                     bool containsSingleAssembly = am.ContainingAssemblySet == null || am.ContainingAssemblySet.AssemblyCount < 2;
 
-                    _context.RunOnAssemblySymbolActions(am.Left?.Element,
+                    context.RunOnAssemblySymbolActions(am.Left?.Element,
                         am.Right[rightIndex]?.Element,
                         am.Left?.MetadataInformation ?? MetadataInformation.DefaultLeft,
                         am.Right[rightIndex]?.MetadataInformation ?? MetadataInformation.DefaultRight,
@@ -57,7 +48,7 @@ namespace Microsoft.DotNet.ApiCompatibility.Rules
                 {
                     if (tm.ShouldDiffElement(rightIndex))
                     {
-                        _context.RunOnTypeSymbolActions(tm.Left,
+                        context.RunOnTypeSymbolActions(tm.Left,
                             tm.Right[rightIndex],
                             tm.ContainingNamespace.ContainingAssembly.Left?.MetadataInformation ?? MetadataInformation.DefaultLeft,
                             tm.ContainingNamespace.ContainingAssembly.Right[rightIndex]?.MetadataInformation ?? MetadataInformation.DefaultRight,
@@ -72,7 +63,7 @@ namespace Microsoft.DotNet.ApiCompatibility.Rules
                         Debug.Assert(mm.ContainingType.Left != null);
                         Debug.Assert(mm.ContainingType.Right[rightIndex] != null);
 
-                        _context.RunOnMemberSymbolActions(
+                        context.RunOnMemberSymbolActions(
                             mm.Left,
                             mm.Right[rightIndex],
                             mm.ContainingType.Left!,
