@@ -6,27 +6,18 @@ namespace Microsoft.DotNet.ApiCompatibility.Runner
     /// <summary>
     /// Options to configure api comparison performed by the <see cref="IApiCompatRunner"/>.
     /// </summary>
-    public readonly struct ApiCompatRunnerOptions
+    public readonly struct ApiCompatRunnerOptions(bool enableStrictMode,
+        bool isBaselineComparison = false)
     {
         /// <summary>
         /// Performs api comparison in strict mode if true.
         /// </summary>
-        public readonly bool EnableStrictMode;
+        public readonly bool EnableStrictMode = enableStrictMode;
 
         /// <summary>
         /// True if assemblies from different roots are compared.
         /// </summary>
-        public readonly bool IsBaselineComparison;
-
-        /// <summary>
-        /// Initializes api compat options
-        /// </summary>
-        public ApiCompatRunnerOptions(bool enableStrictMode,
-            bool isBaselineComparison = false)
-        {
-            EnableStrictMode = enableStrictMode;
-            IsBaselineComparison = isBaselineComparison;
-        }
+        public readonly bool IsBaselineComparison = isBaselineComparison;
 
         /// <inheritdoc />
         public override string ToString() => $"options [ strict mode: {EnableStrictMode}, baseline: {IsBaselineComparison} ]";
