@@ -12,18 +12,9 @@ namespace Microsoft.DotNet.PackageValidation.Validators
     /// Validates that no target framework / rid support is dropped in the latest package.
     /// Reports all the breaking changes in the latest package.
     /// </summary>
-    public class BaselinePackageValidator : IPackageValidator
+    public sealed class BaselinePackageValidator(ISuppressibleLog log,
+        IApiCompatRunner apiCompatRunner) : IPackageValidator
     {
-        private readonly ISuppressibleLog _log;
-        private readonly IApiCompatRunner _apiCompatRunner;
-
-        public BaselinePackageValidator(ISuppressibleLog log,
-            IApiCompatRunner apiCompatRunner)
-        {
-            _log = log;
-            _apiCompatRunner = apiCompatRunner;
-        }
-
         /// <summary>
         /// Validates the latest NuGet package doesn't drop any target framework/rid and does not introduce any breaking changes.
         /// </summary>
@@ -31,7 +22,9 @@ namespace Microsoft.DotNet.PackageValidation.Validators
         public void Validate(PackageValidatorOption options)
         {
             if (options.BaselinePackage is null)
+            {
                 throw new ArgumentNullException(nameof(options.BaselinePackage));
+            }
 
             ApiCompatRunnerOptions apiCompatOptions = new(options.EnableStrictMode, isBaselineComparison: true);
 
@@ -46,14 +39,14 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                     IReadOnlyList<ContentItem>? latestCompileAssets = options.Package.FindBestCompileAssetForFramework(baselineTargetFramework);
                     if (latestCompileAssets == null)
                     {
-                        _log.LogError(new Suppression(DiagnosticIds.TargetFrameworkDropped) { Target = baselineTargetFramework.ToString() },
+                        log.LogError(new Suppression(DiagnosticIds.TargetFrameworkDropped) { Target = baselineTargetFramework.ToString() },
                             DiagnosticIds.TargetFrameworkDropped,
                             string.Format(Resources.MissingTargetFramework,
                                 baselineTargetFramework));
                     }
                     else if (options.EnqueueApiCompatWorkItems)
                     {
-                        _apiCompatRunner.QueueApiCompatFromContentItem(_log,
+                        apiCompatRunner.QueueApiCompatFromContentItem(log,
                             baselineCompileAssets,
                             latestCompileAssets,
                             apiCompatOptions,
@@ -70,14 +63,14 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                     IReadOnlyList<ContentItem>? latestRuntimeAssets = options.Package.FindBestRuntimeAssetForFramework(baselineTargetFramework);
                     if (latestRuntimeAssets == null)
                     {
-                        _log.LogError(new Suppression(DiagnosticIds.TargetFrameworkDropped) { Target = baselineTargetFramework.ToString() },
+                        log.LogError(new Suppression(DiagnosticIds.TargetFrameworkDropped) { Target = baselineTargetFramework.ToString() },
                             DiagnosticIds.TargetFrameworkDropped,
                             string.Format(Resources.MissingTargetFramework,
                                 baselineTargetFramework));
                     }
                     else if (options.EnqueueApiCompatWorkItems)
                     {
-                        _apiCompatRunner.QueueApiCompatFromContentItem(_log,
+                        apiCompatRunner.QueueApiCompatFromContentItem(log,
                             baselineRuntimeAssets,
                             latestRuntimeAssets,
                             apiCompatOptions,
@@ -99,7 +92,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                         IReadOnlyList<ContentItem>? latestRuntimeSpecificAssets = options.Package.FindBestRuntimeAssetForFrameworkAndRuntime(baselineTargetFramework, baselineRuntimeSpecificAssetsRidGroup.Key);
                         if (latestRuntimeSpecificAssets == null)
                         {
-                            _log.LogError(new Suppression(DiagnosticIds.TargetFrameworkAndRidPairDropped) { Target = baselineTargetFramework.ToString() + "-" + baselineRuntimeSpecificAssetsRidGroup.Key },
+                            log.LogError(new Suppression(DiagnosticIds.TargetFrameworkAndRidPairDropped) { Target = baselineTargetFramework.ToString() + "-" + baselineRuntimeSpecificAssetsRidGroup.Key },
                                 DiagnosticIds.TargetFrameworkAndRidPairDropped,
                                 string.Format(Resources.MissingTargetFrameworkAndRid,
                                     baselineTargetFramework,
@@ -107,7 +100,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                         }
                         else if (options.EnqueueApiCompatWorkItems)
                         {
-                            _apiCompatRunner.QueueApiCompatFromContentItem(_log,
+                            apiCompatRunner.QueueApiCompatFromContentItem(log,
                                 baselineRuntimeSpecificAssetsRidGroup.ToArray(),
                                 latestRuntimeSpecificAssets,
                                 apiCompatOptions,
@@ -119,7 +112,9 @@ namespace Microsoft.DotNet.PackageValidation.Validators
             }
 
             if (options.ExecuteApiCompatWorkItems)
-                _apiCompatRunner.ExecuteWorkItems();
+            {
+                apiCompatRunner.ExecuteWorkItems();
+            }
         }
     }
 }
