@@ -12,18 +12,10 @@ namespace Microsoft.DotNet.PackageValidation.Validators
     /// Validates that there are compile time and runtime assets for all the compatible frameworks.
     /// Queues APICompat work items for the applicable compile and runtime assemblies for these frameworks.
     /// </summary>
-    public class CompatibleTfmValidator : IPackageValidator
+    public class CompatibleTfmValidator(ISuppressibleLog log,
+        IApiCompatRunner apiCompatRunner) : IPackageValidator
     {
         private static readonly Dictionary<NuGetFramework, HashSet<NuGetFramework>> s_packageTfmMapping = InitializeTfmMappings();
-        private readonly ISuppressibleLog _log;
-        private readonly IApiCompatRunner _apiCompatRunner;
-
-        public CompatibleTfmValidator(ISuppressibleLog log,
-            IApiCompatRunner apiCompatRunner)
-        {
-            _log = log;
-            _apiCompatRunner = apiCompatRunner;
-        }
 
         /// <summary>
         /// Validates that there are compile time and runtime assets for all the compatible frameworks.
@@ -34,7 +26,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
         {
             ApiCompatRunnerOptions apiCompatOptions = new(options.EnableStrictMode);
 
-            HashSet<NuGetFramework> compatibleTargetFrameworks = new();
+            HashSet<NuGetFramework> compatibleTargetFrameworks = [];
             foreach (NuGetFramework item in options.Package.FrameworksInPackage)
             {
                 compatibleTargetFrameworks.Add(item);
@@ -49,7 +41,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                 IReadOnlyList<ContentItem>? compileTimeAsset = options.Package.FindBestCompileAssetForFramework(framework);
                 if (compileTimeAsset == null)
                 {
-                    _log.LogError(new Suppression(DiagnosticIds.ApplicableCompileTimeAsset) { Target = framework.ToString() },
+                    log.LogError(new Suppression(DiagnosticIds.ApplicableCompileTimeAsset) { Target = framework.ToString() },
                         DiagnosticIds.ApplicableCompileTimeAsset,
                         string.Format(Resources.NoCompatibleCompileTimeAsset,
                             framework));
@@ -59,7 +51,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                 IReadOnlyList<ContentItem>? runtimeAsset = options.Package.FindBestRuntimeAssetForFramework(framework);
                 if (runtimeAsset == null)
                 {
-                    _log.LogError(new Suppression(DiagnosticIds.CompatibleRuntimeRidLessAsset) { Target = framework.ToString() },
+                    log.LogError(new Suppression(DiagnosticIds.CompatibleRuntimeRidLessAsset) { Target = framework.ToString() },
                         DiagnosticIds.CompatibleRuntimeRidLessAsset,
                         string.Format(Resources.NoCompatibleRuntimeAsset,
                             framework));
@@ -67,19 +59,19 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                 // Invoke ApiCompat to compare the compile time asset with the runtime asset if they are not the same assembly.
                 else if (options.EnqueueApiCompatWorkItems)
                 {
-                    _apiCompatRunner.QueueApiCompatFromContentItem(_log,
+                    apiCompatRunner.QueueApiCompatFromContentItem(log,
                         compileTimeAsset,
                         runtimeAsset,
                         apiCompatOptions,
                         options.Package);
                 }
 
-                foreach (string rid in options.Package.Rids.Where(t => IsSupportedRidTargetFrameworkPair(framework, t)))
+                foreach (string rid in options.Package.Rids.Where(packageRid => framework.SupportsRuntimeIdentifier(packageRid)))
                 {
                     IReadOnlyList<ContentItem>? runtimeRidSpecificAsset = options.Package.FindBestRuntimeAssetForFrameworkAndRuntime(framework, rid);
                     if (runtimeRidSpecificAsset == null)
                     {
-                        _log.LogError(new Suppression(DiagnosticIds.CompatibleRuntimeRidSpecificAsset) { Target = framework.ToString() + "-" + rid },
+                        log.LogError(new Suppression(DiagnosticIds.CompatibleRuntimeRidSpecificAsset) { Target = framework.ToString() + "-" + rid },
                             DiagnosticIds.CompatibleRuntimeRidSpecificAsset,
                             string.Format(Resources.NoCompatibleRidSpecificRuntimeAsset,
                                 framework,
@@ -89,7 +81,7 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                     // if the comparison hasn't already happened (when the runtime asset is the same as the runtime specific asset).
                     else if (options.EnqueueApiCompatWorkItems)
                     {
-                        _apiCompatRunner.QueueApiCompatFromContentItem(_log,
+                        apiCompatRunner.QueueApiCompatFromContentItem(log,
                             compileTimeAsset,
                             runtimeRidSpecificAsset,
                             apiCompatOptions,
@@ -99,12 +91,14 @@ namespace Microsoft.DotNet.PackageValidation.Validators
             }
 
             if (options.ExecuteApiCompatWorkItems)
-                _apiCompatRunner.ExecuteWorkItems();
+            {
+                apiCompatRunner.ExecuteWorkItems();
+            }
         }
 
         private static Dictionary<NuGetFramework, HashSet<NuGetFramework>> InitializeTfmMappings()
         {
-            Dictionary<NuGetFramework, HashSet<NuGetFramework>> packageTfmMapping = new();
+            Dictionary<NuGetFramework, HashSet<NuGetFramework>> packageTfmMapping = [];
 
             // creating a map framework in package => frameworks to test based on default compatibilty mapping.
             foreach (OneWayCompatibilityMappingEntry item in DefaultFrameworkMappings.Instance.CompatibilityMappings)
@@ -117,14 +111,11 @@ namespace Microsoft.DotNet.PackageValidation.Validators
                 }
                 else
                 {
-                    packageTfmMapping.Add(forwardTfm, new HashSet<NuGetFramework> { reverseTfm });
+                    packageTfmMapping.Add(forwardTfm, [ reverseTfm ]);
                 }
             }
 
             return packageTfmMapping;
         }
-
-        private static bool IsSupportedRidTargetFrameworkPair(NuGetFramework tfm, string rid) =>
-            tfm.Framework != ".NETFramework" || rid.StartsWith("win", StringComparison.OrdinalIgnoreCase);
     }
 }
