@@ -6,47 +6,38 @@ namespace Microsoft.DotNet.PackageValidation.Validators
     /// <summary>
     /// A package validator option bag that is passed into package validators for configuration.
     /// </summary>
-    public readonly struct PackageValidatorOption
+    /// <remarks>
+    /// Instantiates a new PackageValidatorOption type to be passed into a validator.
+    /// </remarks>
+    public readonly struct PackageValidatorOption(Package package,
+        bool enableStrictMode = false,
+        bool enqueueApiCompatWorkItems = true,
+        bool executeApiCompatWorkItems = true,
+        Package? baselinePackage = null)
     {
         /// <summary>
         /// The latest package that should be validated.
         /// </summary>
-        public Package Package { get; }
+        public Package Package { get; } = package;
 
         /// <summary>
-        /// If true, comparision is performed in strict mode.
+        /// If true, comparison is performed in strict mode.
         /// </summary>
-        public bool EnableStrictMode { get; }
+        public bool EnableStrictMode { get; } = enableStrictMode;
 
         /// <summary>
         /// If true, work items for api compatibility checks are enqueued.
         /// </summary>
-        public bool EnqueueApiCompatWorkItems { get; }
+        public bool EnqueueApiCompatWorkItems { get; } = enqueueApiCompatWorkItems;
 
         /// <summary>
         /// If true, executes enqueued api compatibility work items.
         /// </summary>
-        public bool ExecuteApiCompatWorkItems { get; }
+        public bool ExecuteApiCompatWorkItems { get; } = executeApiCompatWorkItems;
 
         /// <summary>
         /// The baseline package to validate the latest package.
         /// </summary>
-        public Package? BaselinePackage { get; }
-
-        /// <summary>
-        /// Intantiates a new PackageValidatorOption type to be passed into a validator.
-        /// </summary>
-        public PackageValidatorOption(Package package,
-            bool enableStrictMode = false,
-            bool enqueueApiCompatWorkItems = true,
-            bool executeApiCompatWorkItems = true,
-            Package? baselinePackage = null)
-        {
-            Package = package;
-            EnableStrictMode = enableStrictMode;
-            EnqueueApiCompatWorkItems = enqueueApiCompatWorkItems;
-            ExecuteApiCompatWorkItems = executeApiCompatWorkItems;
-            BaselinePackage = baselinePackage;
-        }
+        public Package? BaselinePackage { get; } = baselinePackage;
     }
 }
