@@ -59,9 +59,10 @@ namespace Microsoft.DotNet.GenAPI.Task
         /// </summary>
         public bool IncludeAssemblyAttributes { get; set; }
 
+        /// <inheritdoc />
         protected override void ExecuteCore()
         {
-            GenAPIApp.Run(new MSBuildLog(Log), new GenAPIApp.Context(
+            GenAPIApp.Run(new MSBuildLog(Log),
                 Assemblies!,
                 AssemblyReferences,
                 OutputPath,
@@ -71,7 +72,7 @@ namespace Microsoft.DotNet.GenAPI.Task
                 ExcludeAttributesFiles,
                 RespectInternals,
                 IncludeAssemblyAttributes
-            ));
+            );
         }
     }
 }
