@@ -18,113 +18,51 @@ namespace Microsoft.DotNet.GenAPI
     /// </summary>
     public static class GenAPIApp
     {
-        public readonly struct Context
-        {
-            public Context(
-                string[] assemblies,
-                string[]? assemblyReferences,
-                string? outputPath,
-                string? headerFile,
-                string? exceptionMessage,
-                string[]? excludeApiFiles,
-                string[]? excludeAttributesFiles,
-                bool respectInternals,
-                bool includeAssemblyAttributes)
-            {
-                Assemblies = assemblies;
-                AssemblyReferences = assemblyReferences;
-                OutputPath = outputPath;
-                HeaderFile = headerFile;
-                ExceptionMessage = exceptionMessage;
-                ExcludeApiFiles = excludeApiFiles;
-                ExcludeAttributesFiles = excludeAttributesFiles;
-                RespectInternals = respectInternals;
-                IncludeAssemblyAttributes = includeAssemblyAttributes;
-            }
-
-            /// <summary>
-            /// The path to one or more assemblies or directories with assemblies.
-            /// </summary>
-            public string[] Assemblies { get; }
-
-            /// <summary>
-            /// Paths to assembly references or their underlying directories for a specific target framework in the package.
-            /// </summary>
-            public string[]? AssemblyReferences { get; }
-
-            /// <summary>
-            /// Output path. Default is the console. Can specify an existing directory as well and
-            /// then a file will be created for each assembly with the matching name of the assembly.
-            /// </summary>
-            public string? OutputPath { get; }
-
-            /// <summary>
-            /// Specify a file with an alternate header content to prepend to output.
-            /// </summary>
-            public string? HeaderFile { get; }
-
-            /// <summary>
-            /// Method bodies should throw PlatformNotSupportedException.
-            /// </summary>
-            public string? ExceptionMessage { get; }
-
-            /// <summary>
-            /// The path to one or more api exclusion files with types in DocId format.
-            /// </summary>
-            public string[]? ExcludeApiFiles { get; }
-
-            /// <summary>
-            /// The path to one or more attribute exclusion files with types in DocId format.
-            /// </summary>
-            public string[]? ExcludeAttributesFiles { get; }
-
-            /// <summary>
-            /// If true, includes both internal and public API.
-            /// </summary>
-            public bool RespectInternals { get; }
-
-            /// <summary>
-            /// Includes assembly attributes which are values that provide information about an assembly.
-            /// </summary>
-            public bool IncludeAssemblyAttributes { get; }
-        }
-
         /// <summary>
         /// Initialize and run Roslyn-based GenAPI tool.
         /// </summary>
-        public static void Run(ILog logger, Context context)
+        public static void Run(ILog logger,
+            string[] assemblies,
+            string[]? assemblyReferences,
+            string? outputPath,
+            string? headerFile,
+            string? exceptionMessage,
+            string[]? excludeApiFiles,
+            string[]? excludeAttributesFiles,
+            bool respectInternals,
+            bool includeAssemblyAttributes)
         {
-            bool resolveAssemblyReferences = context.AssemblyReferences?.Length > 0;
+            bool resolveAssemblyReferences = assemblyReferences?.Length > 0;
 
             // Create, configure and execute the assembly loader.
-            AssemblySymbolLoader loader = new(resolveAssemblyReferences, context.RespectInternals);
-            if (context.AssemblyReferences is not null)
+            AssemblySymbolLoader loader = new(resolveAssemblyReferences, respectInternals);
+            if (assemblyReferences is not null)
             {
-                loader.AddReferenceSearchPaths(context.AssemblyReferences);
+                loader.AddReferenceSearchPaths(assemblyReferences);
             }
-            IReadOnlyList<IAssemblySymbol?> assemblySymbols = loader.LoadAssemblies(context.Assemblies);
+            IReadOnlyList<IAssemblySymbol?> assemblySymbols = loader.LoadAssemblies(assemblies);
 
-            string headerFileText = ReadHeaderFile(context.HeaderFile);
+            string headerFileText = ReadHeaderFile(headerFile);
 
             AccessibilitySymbolFilter accessibilitySymbolFilter = new(
-                context.RespectInternals,
+                respectInternals,
                 includeEffectivelyPrivateSymbols: true,
                 includeExplicitInterfaceImplementationSymbols: true);
 
             // Configure the symbol filter
             CompositeSymbolFilter symbolFilter = new();
-            if (context.ExcludeApiFiles is not null)
+            if (excludeApiFiles is not null)
             {
-                symbolFilter.Add(new DocIdSymbolFilter(context.ExcludeApiFiles));
+                symbolFilter.Add(new DocIdSymbolFilter(excludeApiFiles));
             }
             symbolFilter.Add(new ImplicitSymbolFilter());
             symbolFilter.Add(accessibilitySymbolFilter);
 
             // Configure the attribute data symbol filter
             CompositeSymbolFilter attributeDataSymbolFilter = new();
-            if (context.ExcludeAttributesFiles is not null)
+            if (excludeAttributesFiles is not null)
             {
-                attributeDataSymbolFilter.Add(new DocIdSymbolFilter(context.ExcludeAttributesFiles));
+                attributeDataSymbolFilter.Add(new DocIdSymbolFilter(excludeAttributesFiles));
             }
             attributeDataSymbolFilter.Add(accessibilitySymbolFilter);
 
@@ -134,15 +72,15 @@ namespace Microsoft.DotNet.GenAPI
                 if (assemblySymbol is null)
                     continue;
 
-                using TextWriter textWriter = GetTextWriter(context.OutputPath, assemblySymbol.Name);
+                using TextWriter textWriter = GetTextWriter(outputPath, assemblySymbol.Name);
                 textWriter.Write(headerFileText);
 
                 using CSharpFileBuilder fileBuilder = new(logger,
                     symbolFilter,
                     attributeDataSymbolFilter,
                     textWriter,
-                    context.ExceptionMessage,
-                    context.IncludeAssemblyAttributes,
+                    exceptionMessage,
+                    includeAssemblyAttributes,
                     loader.MetadataReferences);
 
                 fileBuilder.WriteAssembly(assemblySymbol);
@@ -165,13 +103,7 @@ namespace Microsoft.DotNet.GenAPI
             }
         }
 
-        /// <summary>
-        /// Creates a TextWriter capable to write into Console or cs file.
-        /// </summary>
-        /// <param name="outputDirPath">Path to a directory where file with `assemblyName`.cs filename needs to be created.
-        ///     If Null - output to Console.Out.</param>
-        /// <param name="assemblyName">Name of an assembly. if outputDirPath is not a Null - represents a file name.</param>
-        /// <returns></returns>
+        // Creates a TextWriter capable of writing into Console or a cs file.
         private static TextWriter GetTextWriter(string? outputDirPath, string assemblyName)
         {
             if (outputDirPath is null)
@@ -188,11 +120,7 @@ namespace Microsoft.DotNet.GenAPI
             return File.CreateText(outputDirPath);
         }
 
-        /// <summary>
-        /// Read the header file if specified, or use default one.
-        /// </summary>
-        /// <param name="headerFile">File with an alternate header content to prepend to output</param>
-        /// <returns></returns>
+        // Read the header file if specified, or use default one.
         private static string ReadHeaderFile(string? headerFile)
         {
             const string defaultFileHeader = """
