@@ -12,11 +12,9 @@ namespace Microsoft.DotNet.GenAPI
 {
     internal static class SyntaxGeneratorExtensions
     {
-        /// <summary>
-        /// Creates a declaration matching an existing symbol.
-        ///     The reason of having this similar to `SyntaxGenerator.Declaration` extension method is that
-        ///     SyntaxGenerator does not generates attributes neither for types, neither for members.
-        /// </summary>
+        // Creates a declaration matching an existing symbol.
+        // The reason of having this similar to `SyntaxGenerator.Declaration` extension method is that
+        // SyntaxGenerator does not generates attributes neither for types, neither for members.
         public static SyntaxNode DeclarationExt(this SyntaxGenerator syntaxGenerator, ISymbol symbol, ISymbolFilter symbolFilter)
         {
             if (symbol.Kind == SymbolKind.NamedType)
@@ -124,7 +122,7 @@ namespace Microsoft.DotNet.GenAPI
             INamedTypeSymbol type,
             ISymbolFilter symbolFilter)
         {
-            List<BaseTypeSyntax> baseTypes = new();
+            List<BaseTypeSyntax> baseTypes = [];
 
             if (type.TypeKind == TypeKind.Class && type.BaseType != null && symbolFilter.Include(type.BaseType))
             {
@@ -144,6 +142,7 @@ namespace Microsoft.DotNet.GenAPI
             baseTypes.AddRange(type.Interfaces
                 .Where(i => symbolFilter.Include(i) && !i.HasInaccessibleTypeArgument(symbolFilter))
                 .Select(i => SyntaxFactory.SimpleBaseType((TypeSyntax)syntaxGenerator.TypeExpression(i))));
+
             return baseTypes.Count > 0 ?
                 SyntaxFactory.BaseList(SyntaxFactory.SeparatedList(baseTypes)) :
                 null;
