@@ -6,32 +6,21 @@ namespace Microsoft.DotNet.ApiSymbolExtensions
     /// <summary>
     /// Class that represents a warning that occurred while trying to load a specific assembly.
     /// </summary>
-    public class AssemblyLoadWarning : IDiagnostic, IEquatable<AssemblyLoadWarning>
+    /// <param name="diagnosticId">String representing the diagnostic ID.</param>
+    /// <param name="referenceId">String representing the ID for the object that the diagnostic was created for.</param>
+    /// <param name="message">String describing the diagnostic.</param>
+    public class AssemblyLoadWarning(string diagnosticId, string referenceId, string message) : IDiagnostic, IEquatable<AssemblyLoadWarning>
     {
         private readonly StringComparer _ordinalComparer = StringComparer.Ordinal;
 
         /// <inheritdoc/>
-        public string DiagnosticId { get; }
+        public string DiagnosticId { get; } = diagnosticId;
 
         /// <inheritdoc/>
-        public string ReferenceId { get; }
+        public string ReferenceId { get; } = referenceId;
 
         /// <inheritdoc/>
-        public string Message { get; }
-
-        /// <summary>
-        /// Creates a new instance of an <see cref="AssemblyLoadWarning"/> class with a given <paramref name="diagnosticId"/>,
-        /// <paramref name="referenceId"/> and <paramref name="message"/>.
-        /// </summary>
-        /// <param name="diagnosticId">String representing the diagnostic ID.</param>
-        /// <param name="referenceId">String representing the ID for the object that the diagnostic was created for.</param>
-        /// <param name="message">String describing the diagnostic.</param>
-        public AssemblyLoadWarning(string diagnosticId, string referenceId, string message)
-        {
-            DiagnosticId = diagnosticId;
-            ReferenceId = referenceId;
-            Message = message;
-        }
+        public string Message { get; } = message;
 
         /// <inheritdoc/>
         public bool Equals(AssemblyLoadWarning? other) => other != null &&
@@ -46,11 +35,15 @@ namespace Microsoft.DotNet.ApiSymbolExtensions
         /// <inheritdoc />
         public override int GetHashCode()
         {
+#if NET
+            return HashCode.Combine(DiagnosticId, ReferenceId, Message);
+#else
             int hashCode = 1447485498;
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(DiagnosticId);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(ReferenceId);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(Message);
             return hashCode;
+#endif
         }
     }
 }
