@@ -164,9 +164,8 @@ namespace Microsoft.DotNet.Cli.New.IntegrationTests
         [InlineData(new object[] { "console", "VB" })]
         public async void AotVariants(string name, string language)
         {
-            // templates have not be updated to 9 yet "net8.0";
-            //string currentDefaultFramework = $"net{Environment.Version.Major}.{Environment.Version.Minor}";
-            string currentDefaultFramework = "net8.0";
+            // template framework needs to be hardcoded here during the major version transition.
+            string currentDefaultFramework = $"net{Environment.Version.Major}.{Environment.Version.Minor}";
 
             string workingDir = CreateTemporaryFolder(folderName: $"{name}-{language}");
             string outputDir = "MyProject";
@@ -336,8 +335,8 @@ namespace Microsoft.DotNet.Cli.New.IntegrationTests
             bool supportsFileScopedNs)
         {
             // Templates have not been updated to net9.0 yet "net8.0";
-            string currentDefaultFramework = "net8.0";
-            //string currentDefaultFramework = $"net{Environment.Version.Major}.{Environment.Version.Minor}";
+            // string currentDefaultFramework = "net8.0";
+            string currentDefaultFramework = $"net{Environment.Version.Major}.{Environment.Version.Minor}";
 
             string workingDir = CreateTemporaryFolder(folderName: $"{name}-{langVersion ?? "null"}-{framework ?? "null"}");
             string outputDir = "MyProject";
